Move AttachmentIdAssignationStrategy to another location
