+++ mailbox/store/src/main/java/org/apache/james/mailbox/store/mail/AttachmentIdAssignationStrategy.java
@@ -1,0 +1,45 @@
+/****************************************************************
+ * Licensed to the Apache Software Foundation (ASF) under one   *
+ * or more contributor license agreements.  See the NOTICE file *
+ * distributed with this work for additional information        *
+ * regarding copyright ownership.  The ASF licenses this file   *
+ * to you under the Apache License, Version 2.0 (the            *
+ * "License"); you may not use this file except in compliance   *
+ * with the License.  You may obtain a copy of the License at   *
+ *                                                              *
+ *   http://www.apache.org/licenses/LICENSE-2.0                 *
+ *                                                              *
+ * Unless required by applicable law or agreed to in writing,   *
+ * software distributed under the License is distributed on an  *
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY       *
+ * KIND, either express or implied.  See the License for the    *
+ * specific language governing permissions and limitations      *
+ * under the License.                                           *
+ ****************************************************************/
+
+package org.apache.james.mailbox.store.mail;
+
+import jakarta.inject.Inject;
+
+import org.apache.james.mailbox.AttachmentIdFactory;
+import org.apache.james.mailbox.model.AttachmentId;
+import org.apache.james.mailbox.model.MessageId;
+import org.apache.james.mailbox.model.ParsedAttachment;
+
+public interface AttachmentIdAssignationStrategy {
+    AttachmentId assign(ParsedAttachment parsedAttachment, MessageId messageId);
+
+    class Default implements AttachmentIdAssignationStrategy {
+        private final AttachmentIdFactory attachmentIdFactory;
+
+        @Inject
+        public Default(AttachmentIdFactory attachmentIdFactory) {
+            this.attachmentIdFactory = attachmentIdFactory;
+        }
+
+        @Override
+        public AttachmentId assign(ParsedAttachment parsedAttachment, MessageId messageId) {
+            return attachmentIdFactory.random();
+        }
+    }
+}
